test data added

+++ src/main/java/ru/trandefil/tm/command/project/ProjectRemoveCommand.java
@@ -1,8 +1,9 @@
 package ru.trandefil.tm.command.project;
 
-import ru.trandefil.tm.api.ServiceLocator;
 import ru.trandefil.tm.api.ProjectService;
+import ru.trandefil.tm.api.ServiceLocator;
 import ru.trandefil.tm.command.AbstractCommand;
+import ru.trandefil.tm.entity.Project;
 import ru.trandefil.tm.service.TerminalService;
 
 import static ru.trandefil.tm.util.UserInputUtil.getNotNullString;
@@ -32,6 +33,11 @@ public class ProjectRemoveCommand extends AbstractCommand {
         final String projectName = getNotNullString(terminalService,
                 "Plz enter project name you want to delete : ");
         final ProjectService projectService = getServiceLocator().getProjectService();
+        final Project project = projectService.getByName(projectName);
+        if (!project.getUser().equals(getServiceLocator().getLoggedUser())) {
+            System.out.println("have no permition to delete this project.");
+            return;
+        }
         projectService.deleteByName(projectName);
     }
